encoder-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
pillow==10.0.1
numpy==1.24.3